Enhance process kill
This change changes a bit the kill behavior.

Now we kill the process group is the main process doesn't
return 0.

And once SIGKILL is sent no need to call Popen.wait().
Popen.wait() can deadlock if a process catch SIGKILL.

Remove usage of os.getpgid that can raise NoSuchProcess.
The pgid is always the pid in our case.