Add Ceph config options for small disks
When running pifpaf on a test machine with
a small amount of disk the deployed Ceph
cluster can permanently enter the HEALT_WARN
state due to warning level being high by default.

This lowers the warning levels and minimum PGs
per OSD so that machines with small disk can be
used.