Start all etcd nodes at the same time
In etcd v3.5, it looks like each node in the static cluster checks if
it can connect to the other nodes while initializing. However, the
current implementation starts nodes one by one. As a result, the first
node waits for the other nodes to be ready forever.

This commit uses threads to start all nodes simultaneously so that each
node can finish the initialize process.